// Frontend/sales-admin/src/app/domain/customers/model/customer-store.dto.ts
export interface CustomerStoreDto {
    id: string;
    firstName: string;
    lastName: string;
    documentNumber: string;
    documentTypeId: string;
    email: string;
    businessName: string;
    ruc: string;
    homePhoneNumber: string;
    cellPhoneNumber: string;
    address: string;
    departmentId: string;
    provinceId: string;
    districtId: string;
    active: boolean;
}